Fixed the ReSharper inspections these changes raised on master
* Dropped two redundant object casts and one redundant long cast
* Qualified a WaitForGraphs doc reference so it resolves from the test class

See TODO-20260822_test_stability.md in pwiz-ai/todos

+++ pwiz_tools/Skyline/Util/UtilInstall.cs
@@ -392,7 +392,7 @@ namespace pwiz.Skyline.Util
         private static bool IsAlreadyExtracted(ZipEntry entry, string installPath)
         {
             var destination = Path.Combine(installPath, entry.FileName.Replace('/', Path.DirectorySeparatorChar));
-            return File.Exists(destination) && new FileInfo(destination).Length == (long) entry.UncompressedSize;
+            return File.Exists(destination) && new FileInfo(destination).Length == entry.UncompressedSize;
         }
     }
